refactor(app): batch2 — extract chrome/link/editor-edit action hooks
